add update message

+++ src/main/java/sweater/repository/MessageRepository.java
@@ -5,6 +5,6 @@ import sweater.domain.Message;
 
 import java.util.List;
 
-public interface MessageRepository extends CrudRepository<Message, Integer> {  //позволяет найти полный список полей, либо найти их по id
+public interface MessageRepository extends CrudRepository<Message, Long> {  //позволяет найти полный список полей, либо найти их по id
     List<Message> findByTag(String tag);
 }
